Added floor, ceil, fac, sin, cos, and tan to lib/math.arc.
Note that sin, cos and tan are currently implemented in terms of scheme primitives rather than being defined by a Taylor series. This could be changed in the future in to remove the dependency on scheme, although care would have to be taken to ensure that the precision and efficiency are comprable to the underlying scheme implementation.