Added lib/dynvars.arc
lib/dynvars is a simple implementation of two styles of dynamic variables: one
with implicit saving of old values in local variables, and one with explicit
saving of old values on a stack. Both support arbitrary settable places as the
"variables" bound to, and are hence suitable for use with thread-local
variables; two simple macros ('dynvar and 'dynvars) to define thread-local
variables are provided.