Revert "Fixed --mute-banner in arc.sh for mz4"; incorrect bugfix (see long desc)
This does eliminate the banner, but it introduces another, more serious,
problem. `-r', which in long form is `--script', causes scheme to run the
provided file without the Scheme REPL underneath it, which means that it is
impossible to drop down into scheme using `:a' or a keyboard interrupt. See
http://arclanguage.org/item?id=10356.

Also, the reverted commit adds another `-q' to the option list, which is
redundant; `-q' is equivalent to `--no-init-file', which is already present.

This reverts commit fb2655b9dfac94265ae81aaa2706b6a1c71491b7.