Removed response cookie parsing. Other minor refactorings.
Response cookie parsing is unnecessary and the previous implementation
was buggy -- it did not support multiple set-cookie lines in the
response.

I removed the functions to give the user consistent expectations -- mkreq and related
utility functions do not assist with interpreting the server response
beyond separating header from body and breaking the header into its
component lines. Response interpretation is entirely left to the user.

Response parsing utilities may still be worthwhile to implement here. A TODO was included to note this.

Thoughts for parsing utilities
--each header line should be tokenized as nested association lists
--keys for status line should be hard-coded as they are not in the
response. all other keys should be read from the response.