experiment: simplify eof-handling
For many years now, reading s-expressions from ports/streams has been
ugly. `read` and friends return nil on eof by default, but there's a
gotcha: if `read` encounters a nil in the stream, it risks being
misinterpreted as eof.

If you think the stream you're reading might contain nils of its own,
`read` lets you override the return value on eof with in a special
symbol. In practice we always use an uninterned symbol for this special
symbol, leading to this verbose idiom:

  (w/uniq eof
    (read stream eof))

After this commit, we always return a special uninterned symbol on eof,
which is saved in the global variable `eof` for comparison. Now you can
just replace the above with:

  (read stream)

However, you need to always compare the result against the value of the
`eof` symbol. Comparing with nil is no longer a valid check for eof.

This change may introduce bugs. If so, it's helping us identify holes in
our test coverage :/

Solution arrived at after discussion with Zachary Kanfer, who ran into
the gotcha for the umpteenth time. Hopefully the last time this trips
somebody up.
  https://github.com/arclanguage/anarki/issues/61

We've basically reinvented Racket's eof-object? API, which was
conspicuous in ac.scm if we had but looked with eyes open.

I considered switching Racket-based primitives like `readb` and `readc`
as well to return the global eof symbol on end-of-file. But that seems
like a foolish consistency considering they can't ever return symbols in
the happy path.