keep atstrings off by default
They're confusing to newcomers.

If we were doing a 100-year language from scratch we'd probably decide
atstrings weren't worth the hassle. But we're stuck with it in news.arc
until somebody invests the effort to tear it out, so we'll scope its use
tightly. It looks like we can just turn on atstrings around its uses in
news.arc, and the news app continues to work fine, as shown by this example:

  $ ./arc.sh
  arc> (declare 'atstrings nil)
  arc> (top-right "abcd" "/")  ; assuming we have a user called "abcd"
  ; result doesn't contain "@(karma user)"

Let's see if we uncover subtle bugs.

The other place we currently use atstrings (besides news.arc) is our automated
tests. This is because unit-test.arc doesn't currently support running
setup/teardown code around a suite.