bugfix in pretty-printing
Caused by my dropping support for dotted lists in `len` on Jun 23.

In the process, also upgrade lib/ppr.arc.t to unit-test.arc v1.0.

I don't know why we had that extra check in the handling of custom
macros, and we have no tests for it. So I'm removing it and risking
further regression. At least we'll then know what test to write for it.