Fix breakage in `nappend`
It had been broken since commit 95ff0937 in Nov 2017, when I moved it from
a library into arc.arc. Unfortunately I moved it to the wrong place, before
assignment was defined. Not only did it break `nappend`, it also yielded
cryptic error messages when attempting to use `nappend`.

`nappend` was originally introduced because I wanted it for spliceable
lists. Unfortunately I never wrote tests for it running on regular lists,
hiding the bug for over a year. Now starting to add tests.

Thanks Kinnard Hockenhull for reporting this issue.