fix breakage in news app
Fixes #175.

In the process I've temporarily tossed out appdir* and everything that
used it, just to unblock a user. This included the ability to create new
skeleton app directories, because it wasn't clear to me what directory
app-create is intended to be run from.

This commit is absolutely a temporary change and not intended to
indicate that these features shouldn't exist. I'd love if somebody
brought them back in a working form. But I'd rather have few features
that work than lots of features that don't work.

The news app is the #1 thing newcomers try out in this repo. It should
always work. Anytime we modify our app infrastructure we should always
manually test that apps/news/run-news works.